round timer, end leaderboard, standardized answer buttons
Added a timer for each round. When new round starts timer disappears on old question.
Show leaderboard at end of the game now.
Some answers were too long for buttons which would cause an API error with discord since discord will not accept anything over 80 characters for a label on a button, so i made the answers a text field inside the question embed, which standardized the buttons anyways.
Changed wait utility file to a reusable.js utility file which i plan to use to house all the random functions that i think have reusable value
Experimented a bit more with buttons on display components (test.js), and i believe there is a good chance i switch to using display components as opposed to embeds for the questions each round