Finished adding options to the api call and handling & FIXED DUPLICATE EVENT EMISSIONS
Added options to slash command that lets the user customize their games. Most of the options are handled (all the API queries) but the option that changes the game mode to a 'first till x points' is not yet working.

ALSO fixed this bug where when i create multiple instances of a game (aka when the command is run concurrently in different channels/servers) they all listen to event emissions of the same name, hence i added an instance counter that is stuck to the end of emissions to tell them all apart.